# back/docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /equipment:
    get:
      operationId: EquipmentController_findAll
      parameters: []
      responses:
        '200':
          description: List of equipment
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                      example: 10
                    name:
                      type: string
                      example: Longsword
                    equipmentTypeId:
                      type: number
                      example: 7
                    cost:
                      type: number
                      example: 100
                    maxDurability:
                      type: number
                      example: 50
                    currentDurability:
                      type: number
                      example: 50
      security:
        - bearer: []
      tags:
        - Equipment
    post:
      operationId: EquipmentController_create
      parameters: []
      requestBody:
        required: true
        description: New equipment payload
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Longsword
                equipmentTypeId:
                  type: number
                  example: 7
                cost:
                  type: number
                  example: 100
                maxDurability:
                  type: number
                  example: 50
              required:
                - name
                - equipmentTypeId
                - cost
                - maxDurability
              additionalProperties: false
      responses:
        '201':
          description: Equipment created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 10
                  name:
                    type: string
                    example: Longsword
                  equipmentTypeId:
                    type: number
                    example: 7
                  cost:
                    type: number
                    example: 100
                  maxDurability:
                    type: number
                    example: 50
                  currentDurability:
                    type: number
                    example: 50
      security:
        - bearer: []
      tags:
        - Equipment
  /equipment/{id}:
    get:
      operationId: EquipmentController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Equipment ID
          schema:
            example: 10
            type: number
      responses:
        '200':
          description: Equipment by id
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 10
                  name:
                    type: string
                    example: Longsword
                  equipmentTypeId:
                    type: number
                    example: 7
                  cost:
                    type: number
                    example: 100
                  maxDurability:
                    type: number
                    example: 50
                  currentDurability:
                    type: number
                    example: 50
      security:
        - bearer: []
      tags:
        - Equipment
    patch:
      operationId: EquipmentController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Equipment ID
          schema:
            example: 10
            type: number
      requestBody:
        required: true
        description: Fields to update (partial)
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Enchanted Longsword
                equipmentTypeId:
                  type: number
                  example: 8
                cost:
                  type: number
                  example: 150
                maxDurability:
                  type: number
                  example: 60
              additionalProperties: false
      responses:
        '200':
          description: Updated equipment
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 10
                  name:
                    type: string
                    example: Enchanted Longsword
                  equipmentTypeId:
                    type: number
                    example: 8
                  cost:
                    type: number
                    example: 150
                  maxDurability:
                    type: number
                    example: 60
                  currentDurability:
                    type: number
                    example: 45
      security:
        - bearer: []
      tags:
        - Equipment
    delete:
      operationId: EquipmentController_delete
      parameters:
        - name: id
          required: true
          in: path
          description: Equipment ID
          schema:
            example: 10
            type: number
      responses:
        '200':
          description: Delete result
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 10
                  deleted:
                    type: boolean
                    example: true
      security:
        - bearer: []
      tags:
        - Equipment
  /equipment-stocks:
    get:
      operationId: EquipmentStocksController_findAll
      parameters: []
      responses:
        '200':
          description: List of equipment stocks
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                      example: 1
                    equipmentId:
                      type: number
                      example: 12
                    quantity:
                      type: number
                      example: 25
                    available:
                      type: number
                      example: 20
                    createdAt:
                      type: string
                      format: date-time
                      example: '2025-10-30T12:00:00.000Z'
                    updatedAt:
                      type: string
                      format: date-time
                      example: '2025-10-30T12:34:56.000Z'
      security:
        - bearer: []
      tags:
        - Equipment Stocks
    post:
      operationId: EquipmentStocksController_create
      parameters: []
      requestBody:
        required: true
        description: New equipment stock payload
        content:
          application/json:
            schema:
              type: object
              properties:
                equipmentId:
                  type: number
                  example: 12
                quantity:
                  type: number
                  example: 25
              required:
                - equipmentId
              additionalProperties: false
      responses:
        '201':
          description: Equipment stock created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 3
                  equipmentId:
                    type: number
                    example: 12
                  quantity:
                    type: number
                    example: 25
                  available:
                    type: number
                    example: 25
                  createdAt:
                    type: string
                    format: date-time
                    example: '2025-10-30T12:00:00.000Z'
                  updatedAt:
                    type: string
                    format: date-time
                    example: '2025-10-30T12:00:00.000Z'
      security:
        - bearer: []
      tags:
        - Equipment Stocks
  /equipment-stocks/{id}:
    get:
      operationId: EquipmentStocksController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Equipment stock ID
          schema:
            example: 1
            type: number
      responses:
        '200':
          description: Equipment stock by id
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 1
                  equipmentId:
                    type: number
                    example: 12
                  quantity:
                    type: number
                    example: 25
                  available:
                    type: number
                    example: 20
                  createdAt:
                    type: string
                    format: date-time
                    example: '2025-10-30T12:00:00.000Z'
                  updatedAt:
                    type: string
                    format: date-time
                    example: '2025-10-30T12:34:56.000Z'
      security:
        - bearer: []
      tags:
        - Equipment Stocks
    patch:
      operationId: EquipmentStocksController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Equipment stock ID
          schema:
            example: 3
            type: number
      requestBody:
        required: true
        description: Fields to update (partial)
        content:
          application/json:
            schema:
              type: object
              properties:
                equipmentId:
                  type: number
                  example: 15
                quantity:
                  type: number
                  example: 30
              additionalProperties: false
      responses:
        '200':
          description: Updated equipment stock
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 3
                  equipmentId:
                    type: number
                    example: 15
                  quantity:
                    type: number
                    example: 30
                  available:
                    type: number
                    example: 28
                  createdAt:
                    type: string
                    format: date-time
                    example: '2025-10-30T12:00:00.000Z'
                  updatedAt:
                    type: string
                    format: date-time
                    example: '2025-10-30T12:45:00.000Z'
      security:
        - bearer: []
      tags:
        - Equipment Stocks
    delete:
      operationId: EquipmentStocksController_delete
      parameters:
        - name: id
          required: true
          in: path
          description: Equipment stock ID
          schema:
            example: 3
            type: number
      responses:
        '200':
          description: Delete result
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 3
                  deleted:
                    type: boolean
                    example: true
      security:
        - bearer: []
      tags:
        - Equipment Stocks
...
info:
  title: API "Les kips AH"
  description: |-
    This is the API information regarding the project "Les capuches d'Opale".

    C'est un outil de gestion d'aventurier pour une guilde.
  version: 1.0.0
  contact: {}
  termsOfService: https://swagger.io/terms/
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http